Uptade sv1 client message handler to return a general message
IsClient::message_handler should return a general message and not a
response cause when the client is initializing and response from the
server are received we need to return a sv1 message and not a response.
Eg when the configure response arrive from the server the client have to
generate and send a subscribe message.